# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bhaleri Online API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /buses:
    get:
      operationId: listBuses
      tags: [buses]
      summary: List all bus routes
      responses:
        "200":
          description: List of bus routes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bus"
    post:
      operationId: createBus
      tags: [buses]
      summary: Add a bus route (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bus"
  /buses/{id}:
    patch:
      operationId: updateBus
      tags: [buses]
      summary: Update bus route (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bus"
    delete:
      operationId: deleteBus
      tags: [buses]
      summary: Delete bus route (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List job listings
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Post a job (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete job (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /emergency:
    get:
      operationId: listEmergency
      tags: [emergency]
      summary: List emergency contacts
      responses:
        "200":
          description: List of emergency contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmergencyContact"
    post:
      operationId: createEmergency
      tags: [emergency]
      summary: Add emergency contact (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmergencyContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyContact"
  /emergency/{id}:
    patch:
      operationId: updateEmergency
      tags: [emergency]
      summary: Update emergency contact (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmergencyContactUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyContact"
    delete:
      operationId: deleteEmergency
      tags: [emergency]
      summary: Delete emergency contact (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
...
    Bus:
      type: object
      required: [id, name, route, departureTime, fare, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        route:
          type: string
        departureTime:
          type: string
        arrivalTime:
          type: ["string", "null"]
        fare:
          type: string
        createdAt:
          type: string
    BusInput:
      type: object
      required: [name, route, departureTime, fare]
      properties:
        name:
          type: string
        route:
          type: string
        departureTime:
          type: string
        arrivalTime:
          type: ["string", "null"]
        fare:
          type: string
    BusUpdate:
      type: object
      properties:
        name:
          type: string
        route:
          type: string
        departureTime:
          type: string
        arrivalTime:
          type: ["string", "null"]
        fare:
          type: string
    Job:
      type: object
      required: [id, title, contactNumber, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        contactNumber:
          type: string
        salary:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        createdAt:
          type: string
    JobInput:
      type: object
      required: [title, contactNumber]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        contactNumber:
          type: string
        salary:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        contactNumber:
          type: string
        salary:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
    Event:
      type: object
      required: [id, title, date, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
        time:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        createdAt:
          type: string
    EventInput:
      type: object
      required: [title, date]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
        time:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
        time:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
...
    EmergencyContact:
      type: object
      required: [id, name, contactNumber, category, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        contactNumber:
          type: string
        category:
          type: string
          enum: [hospital, ambulance, police, electricity, fire, other]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    EmergencyContactInput:
      type: object
      required: [name, contactNumber, category]
      properties:
        name:
          type: string
        contactNumber:
          type: string
        category:
          type: string
          enum: [hospital, ambulance, police, electricity, fire, other]
        description:
          type: ["string", "null"]
    EmergencyContactUpdate:
      type: object
      properties:
        name:
          type: string
        contactNumber:
          type: string
        category:
          type: string
          enum: [hospital, ambulance, police, electricity, fire, other]
        description:
          type: ["string", "null"]
...